recommend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Bottleneck==1.3.7
category-encoders==1.3.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cornac==1.16.0
cycler==0.11.0
Django==4.2.4
djangorestframework==3.14.0
exceptiongroup==1.1.3
filelock==3.12.3
fonttools==4.42.1
fsspec==2023.9.0
gunicorn==21.2.0
huggingface-hub==0.16.4
hypothesis==6.84.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.0.3
joblib==1.3.2
kiwisolver==1.4.5
lightfm==1.17
lightgbm==4.0.0
llvmlite==0.40.1
MarkupSafe==2.1.3
matplotlib==3.7.2
memory-profiler==0.61.0
mpmath==1.3.0
multimethod==1.9.1
mypy-extensions==1.0.0
nltk==3.8.1
numba==0.57.1
numpy==1.24.4
packaging==23.1
pandas==1.5.3
pandera==0.16.1
patsy==0.5.3
Pillow==10.0.0
powerlaw==1.5
psutil==5.9.5
psycopg==3.1.10
psycopg-binary==3.1.10
pydantic==1.10.12
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
recommenders==1.1.1
regex==2023.8.8
requests==2.31.0
retrying==1.3.4
safetensors==0.3.3
scikit-learn==1.0.2
scikit-surprise==1.1.3
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
statsmodels==0.14.0
threadpoolctl==3.2.0
tokenizers==0.13.3
tqdm==4.66.1
transformers==4.33.1
typeguard==4.1.4
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wrapt==1.15.0
zipp==3.16.2
